% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{buildPseudoTimeSeries}
\alias{buildPseudoTimeSeries}
\title{Create a pseudo-time-seires}
\usage{
buildPseudoTimeSeries(cds, branch_points = NULL)
}
\arguments{
\item{cds}{A cds which has been ordered with Monocle 2}

\item{branch_points}{Vector for the branch points. If it is null, there is no branching in the data}
}
\value{
a list storing the pseudo-time-series data and a vector storing the run id for each cell (row) in the data
}
\description{
This function takes a cds (branched or not) and then convert to a pseudotime series.
}
\examples{
\dontrun{
lung <- load_lung() 
lung_data_list <- buildPseudoTimeSeries(lung, branch_points = 1)
}
}
