% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_estimators.R
\name{cal_cross_map}
\alias{cal_cross_map}
\title{Perform cross-convergence mapping test for two genes.}
\usage{
cal_cross_map(ordered_exprs_mat, lib_colum, target_column, RNGseed = 2016)
}
\arguments{
\item{ordered_exprs_mat}{CellDataSet for the experiment}

\item{lib_colum}{The index (or name) of the column to cross map from (same as in rEDM).}

\item{target_column}{The index (or name) of the column to cross map to (same as in rEDM).}

\item{RNGseed}{will set a seed for the random number generator, enabling reproducible runs of ccm with randomly generated libraries (same as in rEDM).}
}
\value{
a
}
\description{
This function uses the CCM function from rEDM package to determine if one time series contain the necessary 
dynamic information to recover the influence of another causal variable. See the \package{rEDM} for details
of the ccm function.
}
\examples{
\dontrun{
lung <- load_lung() 
gene_pairs_mat <- matrix(c('H19', 'Ccnd2', 'Ccnd2', 'Scnn1g'), ncol = 2)
cal_cross_map(lung, gene_pairs_mat)
}
}
