% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_estimators.R
\name{cal_grangertest}
\alias{cal_grangertest}
\title{Perform the Granger causality test for two genes.}
\usage{
cal_grangertest(ordered_exprs_mat)
}
\arguments{
\item{ordered_exprs_mat}{CellDataSet for the experiment}
}
\value{
a matrix
}
\description{
This function uses the grangertest function from lmtest package to determine if one time series contains the necessary 
dynamic information to recover the influence of another causal variable. Note that granger casuality only deals with linear relationships. 
See the \package{lmtest} for details
of the grangertest function.
}
\examples{
\dontrun{
lung <- load_lung() 
gene_pairs_mat <- matrix(c('H19', 'Ccnd2', 'Ccnd2', 'Scnn1g'), ncol = 2)
cal_grangertest(lung, gene_pairs_mat)
}
}
