% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_estimators.R
\name{cal_grangertest_var}
\alias{cal_grangertest_var}
\title{Perform the Granger causality test for two genes.}
\usage{
cal_grangertest_var(ordered_exprs_mat)
}
\arguments{
\item{ordered_exprs_mat}{CellDataSet for the experiment}
}
\value{
a ggplot2 plot object
}
\description{
This function uses the grangertest function from vars package to determine if one time series contain the necessary 
dynamic information to recover the influence of another causal variable. See the \package{vars} for details
of the VAR or causality function.
}
\examples{
\dontrun{
lung <- load_lung() 
gene_pairs_mat <- matrix(c('H19', 'Ccnd2', 'Ccnd2', 'Scnn1g'), ncol = 2)
cal_cross_map(lung, gene_pairs_mat)
}
}
