% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_estimators.R
\name{cal_knn_mi}
\alias{cal_knn_mi}
\title{Calculate the mutual information between all pairs of genes using the KSG estimator for MI.}
\usage{
cal_knn_mi(cds, k = 5)
}
\arguments{
\item{cds}{CellDataSet for the experiment, columns ordered by exprimental time or pseudotime with Monocle 2}

\item{k}{Number of nearest neighbors used for calculating the mutual information values.}
}
\value{
a list for the mutual information values for all possible pairs of genes.
}
\description{
This function calculates the mutual information using the KSG estimators.
}
\examples{
\dontrun{
lung <- load_lung() 
cal_knn_mi(lung)
}
}
