% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_estimators.R
\name{cal_knn_mi_parmigene}
\alias{cal_knn_mi_parmigene}
\title{Calculate the mutual information between all gene pairs}
\usage{
cal_knn_mi_parmigene(cds)
}
\arguments{
\item{cds}{CellDataSet for the experiment}
}
\value{
a ggplot2 plot object
}
\description{
This function uses the knnmi.all function from \package{parmigene} package to calculate the mutual information between all pairs of genes in the data. 
See the \package{parmigene} for details of the knnmi.all function.
}
\examples{
\dontrun{
lung <- load_lung() 
cal_knn_mi_parmigene(t(exprs(lung))
}
}
