% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{calculate_conditioned_rdi}
\alias{calculate_conditioned_rdi}
\title{Calculate conditionally RDI value}
\usage{
calculate_conditioned_rdi(genes_data, super_graph = NULL, rdi_list,
  top_incoming_k = 1)
}
\arguments{
\item{genes_data}{Time / Pseudotime series expression data (Rows are samples while columns are features).}

\item{super_graph}{A graph including all possible interactions used for performing the causality inference. When it is NULL, 
all possible pairs of interactions for the genes_data are used, otherwise only the interactions specified in the graph are used.
Note that the super_graph only accept integer matrix for now (each integer corresponds to a particular gene in the genes_data).}

\item{rdi_list}{a list returned from calculate_rdi function.}

\item{top_incoming_k}{The number of genes to be conditioned on when calculating the conditional RDI values}

\item{genes_data}{pseudo-time/time series for the gene expression data}
}
\value{
a dataframe storing conditional RDI results. First two columns are the id names for the genes.
Third column is the conditional RDI value.
}
\description{
This function estimates the RDI value for all gene-pair combination from the genes_data (or that specified in the super_graph) 
in the pseudotime/time series data, conditioned on the top top_incoming_k incoming nodes. 
The result from the calculate_rdi function will be used to identify the proper delay for the gene-pair under test corresponds to that 
with highest RDI as well as the proper delays for the k incoming nodes which corresponds to that with highest RDI values.
}
