% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_conditioned_rdi_multiple_run_wrap}
\alias{calculate_conditioned_rdi_multiple_run_wrap}
\title{calculate_conditioned_rdi_multiple_run_wrap}
\usage{
calculate_conditioned_rdi_multiple_run_wrap(expr_data, super_graph,
  max_rdi_value, max_rdi_delays, run_vec, k, uniformalize)
}
\arguments{
\item{expr_data}{a matrix for all variables in a time-series where each row is a time point and each column is a variable (for example, a gene). 
The rows are ordered according to time, from earliest to latest.}

\item{super_graph}{An integer matrix where each row is the variable IDs or gene pairs you would like to estimate restricted direct information (RDI) from.}

\item{max_rdi_value}{A numeric matrix where each element corresponding to the maximal (identified from a series of time lags when calculating the rdi values) rdi value between two variables.}

\item{max_rdi_delays}{An integer matrix where each element corresponding to the time delay corresponding to the maximal (identified from a series of time lags when calculating the rdi values) rdi value between two variables.}

\item{run_vec}{An integer vector keeping the run id for each sample.}

\item{k}{An integer for the number of incoming nodes to be conditioned on.}

\item{uniformalize}{Whether or not you want to use ucmi to calculate rdi. Default to be false.}
}
\value{
a numeric matrix of conditional RDI values for all possible pairs of genes from expr_data. If the gene pairs is not encoded in the super_graph, it will remain as 0 in the matrix.
}
\description{
This function estimates the conditional DIRECTED mutual information for all gene pairs in the expr_data matrix when you have a SINGLE run of the processes.
}
\details{
\code{calculate_conditioned_rdi_multiple_run_wrap}, similar to calculate_conditioned_rdi_cpp_wrap, takes an expression matrix (expr_data) and possible gene pairs (encoded in super_graph), as well as the matrices of maximal rdi value or the delays corresponding to those values to 
estimate the conditional restricted direct information, conditioning on top k incoming nodes. It, however, differs to calculate_conditioned_rdi_cpp_wrap, in that it can concatenate different experiments (runs) into a single run of the data for causality estimation.
}
