% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{calculate_rdi}
\alias{calculate_rdi}
\title{Calculate RDI values}
\usage{
calculate_rdi(genes_data, delays, super_graph = NULL, turning_points = NULL,
  method = 1)
}
\arguments{
\item{genes_data}{Time / Pseudotime series expression data (Rows are samples while columns are features).}

\item{delays}{A vector of time delays used during information transfer estimation between genes}

\item{super_graph}{A graph including all possible interactions used for performing the causality inference. When it is NULL, 
all possible pairs of interactions for the genes_data are used, otherwise only the interactions specified in the graph are used.
Note that the super_graph only accept integer matrix for now (each integer corresponds to a particular gene in the genes_data).}

\item{turning_points}{pseudo-time/time series for the gene expression data}

\item{method}{a list returned from calculate_rdi function.}
}
\value{
a dataframe storing conditional RDI results. First two columns are the id names for the genes.
}
\description{
This function estimates the RDI value for all gene-pair combination from the genes_data (or that specified in the super_graph) 
in the pseudotime/time series data.
}
\examples{
\dontrun{
lung <- load_lung() 
lung_update <- estimate_turning_point(lung, type = 'value')
}
}
