% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_rdi_cpp_wrap}
\alias{calculate_rdi_cpp_wrap}
\title{calculate_rdi_cpp_wrap}
\usage{
calculate_rdi_cpp_wrap(expr_data, delays, super_graph, turning_points, method,
  uniformalize)
}
\arguments{
\item{expr_data}{a matrix for all variables in a time-series where each row is a time point and each column is a variable (for example, a gene). 
The rows are ordered according to time, from earliest to latest.}

\item{delays}{An integer vector storing the time delays between pairs of variables you would like to try.}

\item{super_graph}{An integer matrix where each row is the variable IDs or gene pairs you would like to estimate restricted direct information (RDI) from.}

\item{turning_points}{Either 0 or a numeric vector describing the inflection point (linear trajectory) or branch point (bifurcation point) for each gene. 
If the turning_point for each gene is provided, the time delay will be estimated based on the turning point.}

\item{method}{An integer of either 1 or 2 to determine which information metric will be used to quantify the causality. 
If method is 1, then lagged mutual information will be used; if method is 2, then the restricted direct information will be used.}

\item{uniformalize}{Whether or not you want to use ucmi to calculate rdi. Default to be false.}
}
\value{
a numeric matrix of conditional RDI values for all possible pairs of genes from expr_data. If the gene pairs is not encoded in the super_graph, it will remain as 0 in the matrix.
}
\description{
This function estimates the DIRECTED mutual information for all gene pairs in the expr_data matrix when you have a SINGLE run of the processes
}
\details{
\code{calculate_rdi_cpp_wrap} takes an expression matrix (expr_data) and possible gene pairs (encoded in super_graph) to estimate the restricted 
direct information based on the time delay which can be estimated from the turning_point vector.
}
