% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{calculate_temporal_causality}
\alias{calculate_temporal_causality}
\title{Calculate temporal RDI or conditional RDI (cRDI) values.}
\usage{
calculate_temporal_causality(cds_subset, super_graph = NULL,
  conditioning = FALSE, window_size = 40, delay = 1, smoothing = TRUE,
  span = 0.1, uniformalize = FALSE, log = TRUE, pseudo_cnt = 1,
  verbose = FALSE, ...)
}
\arguments{
\item{cds_subset}{A cds_subset which has been ordered with Monocle 2. Currently it only deals with a linear trajectory.}

\item{super_graph}{A matrix for the valid gene-pairs we will consider for inferring the temporal RDI or cRDI.}

\item{conditioning}{A logic argument to determine whether or not the conditional RDI will be calculated.}

\item{window_size}{The window size we will use for calculating the temporal RDI.}

\item{delay}{The time delay used in calculating the RDI or cRDI values.}

\item{smoothing}{A logic argument to determine whether or not we will first apply loess smoothing before we calculate the temporal RDI. Default to be TRUE.}

\item{span}{The parameter α which controls the degree of smoothing, passed into the loess functioin. Default is set to be 0.1.}

\item{uniformalize}{Whether or not you want to use ucmi to calculate rdi. Default to be false.}

\item{verbose}{A logical argument to determine whether or not we will print the detailed running information.}
}
\value{
A list of two arrays where the first one is the temporal RDI value and second one is the conditional RDI values. In each array, the first 
dimension corresponds to the sliding window, the second and third dimensions correspond to the genes.
}
\description{
This function is developed to calculate the temporal RDI or cRDI which can reveal the time point where the regulator will affect the target gene.
}
\examples{
\dontrun{
lung <- load_lung() 
plot_cell_trajectory(lung)
plot_cell_trajectory(lung, color_by = 'Time')
lung_AT1_lineage <- lung[, pData(lung)$State \%in\% c(1, 3)]
temp_causality_list <- calculate_temporal_causality(lung_AT1_lineage)
}
}
