% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{calculate_umi}
\alias{calculate_umi}
\title{Calculate uMI values}
\usage{
calculate_umi(cds_subset, super_graph = NULL, log = TRUE, pseudo_cnt = 1,
  k = 5L, method = 1L, k_density = 5L, bw = 0)
}
\arguments{
\item{cds_subset}{A cds_subset which has been ordered with Monocle 2}

\item{super_graph}{A graph including all possible interactions used for performing the causality inference. When it is NULL, 
all possible pairs of interactions for the genes_data are used, otherwise only the interactions specified in the graph are used.
Note that the super_graph only accept integer matrix for now (each integer corresponds to a particular gene in the genes_data).}

\item{k}{Number for nearest neighbors used in entropy calculation.}

\item{k_density}{The number of k nearest neighbors you would like to use when calculating the density (only applicable when method == 2 or using knn based density estimation).}

\item{bw}{Bindwidth used for the kernel density estimator. Currently it is not used. The bindwidth in the kde function is automatically estimated.}

\item{methodWhich}{2D density estimator you would like to use. 1 is kde estimator and 2 is knn based estimator. Default to be 1.}
}
\value{
a numeric matrix storing a uMI result for all gene-pair combination from the genes_data (or that specified in the super_graph).
}
\description{
This function estimates the uniformed mutual information value for all gene-pair combination from the genes_data (or that specified in the super_graph)
}
\examples{
\dontrun{
lung <- load_lung() 
AT1_lung <- lung[, pData(lung)$State \%in\% c(2, 3)]
rdi_res <- calculate_umi(AT1_lung, delays = 10, method = 1)
}
}
