% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmi}
\alias{cmi}
\title{cmi}
\usage{
cmi(x, y, z, k, normalize)
}
\arguments{
\item{x}{one random variable from the time-series data}

\item{y}{another random variable from the time-series data}

\item{z}{condition random variable for variables (x, y) from the time-series data}

\item{k}{number for nearest neighbors used in entropy calculation}

\item{normalize}{A logic flag to determine whether or not you want to normalize the MI value by the density of x.}
}
\value{
a estimated conditional mutual information value between two variables (x, y), conditioning on a third variable z.
}
\description{
This subroutine calculates the volume of a d-dimensional unit ball for Euclidean norm
}
\details{
\code{cmi} takes two random variable x and y and estimated their mutual information conditioned on the third random variable z
using the KSG estimator. 
It relies on the ANN package to query the kNN with KDTree algorithm.
}
