% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command-args.R
\name{command_args}
\alias{command_args}
\title{Command line arguments}
\usage{
command_args(
  x = commandArgs(trailingOnly = TRUE),
  include = getOption("scribe.include", c("help", "version", NA_character_)),
  super = include,
  string
)
}
\arguments{
\item{x}{Command line arguments as a character vector.  Use \code{I()} to pass a
single (combined) string of arguments.}

\item{include}{Special default arguments to included.  See \verb{$initialize()} in
\link{scribeCommandArgs} for more details.}

\item{super}{When \code{TRUE} the \link{scribeCommandArgs} object will be initialized
with standard \emph{super} arguments (e.g., \code{---help}, \code{---version})}

\item{string}{Deprecated.  Use \code{command_args(I())} to pass a single string of
arguments.}
}
\value{
A \link{scribeCommandArgs} object
}
\description{
Make a new \link{scribeCommandArgs} object
}
\examples{
command_args()
command_args(c("-a", 1, "-b", 2))
command_args(I("-a 1 -b 2"))
}
\seealso{
Other scribe:
\code{\link[=new_arg]{new_arg()}},
\code{\link{scribeArg-class}},
\code{\link{scribeCommandArgs-class}}
}
\concept{scribe}
