% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{create_super_graph}
\alias{create_super_graph}
\title{Calculate temporal RDI or conditional RDI (cRDI) values.}
\usage{
create_super_graph(cds, TFs = NULL, informative_genes = NULL)
}
\arguments{
\item{cds}{A cds which contains the genes from the TFs and informative_genes vectors in the gene_short_name column of the cds's pData.}

\item{TFs}{A vector for the gene short name of the TFs, default to be NULL (All genes in the cds will be used).}

\item{informative_genes}{A vector for the gene short name of the informative genes, default to be NULL (All genes in the cds will be used).}
}
\value{
A character matrix where the first column corresponds to the regulators and the second column corresponds to the targets.
}
\description{
This function is developed to calculate the temporal RDI or cRDI which can reveal the time point where the regulator will affect the target gene.
}
\examples{
\dontrun{
lung <- load_lung() 
super_graph <- create_super_graph(lung)
}
}
