% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{di_single_run_conditioned}
\alias{di_single_run_conditioned}
\title{di_single_run_conditioned}
\usage{
di_single_run_conditioned(x, y, z, n, uniformalize)
}
\arguments{
\item{x}{one random variable from the time-series data}

\item{y}{another random variable from the time-series data}

\item{z}{z is a dataframe (or matrix) containing the data of other processes upon the past of which the mi is conditioned}

\item{n}{Parameter n determines the the number of previous time samples upon which the mi is conditioned}

\item{uniformalize}{Whether or not you want to use ucmi to calculate rdi. Default to be false.}
}
\value{
a numeric value for the condition mutual information estimator variables (x, y), conditioning on a third variable, z.
}
\description{
This function estimates the CONDITIONED DIRECTED mutual information from X to Y, conditioning on a third variable, z, when you have a SINGLE run of the processes
}
\details{
\code{di_single_run_conditioned} takes two random variables x, y and z as well as the parameter n to calculate the direct information conditioned on variable z.
}
