% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{dn_sparsifer}
\alias{dn_sparsifer}
\title{Sparsify the direct network with biologically inspired constraints}
\usage{
dn_sparsifer(W, alpha = 2, lambda = 20, B = 184)
}
\arguments{
\item{W}{A matrix which contains the causality score between genes. This matrix is asymmetric.}

\item{lambda}{The parameter of the exponential distribution.}

\item{B}{The budget parameter introduced to prevent trivial solution.}

\item{alphae}{The exponent of the power law distribution.}
}
\value{
A new asymmeric matrix which contains a sparse direct network structure estimated based on the network sparsifier.
}
\description{
This function implements the direct network sparsifier algorithm described in the supplementary file of Scribe manuscript. 
It relies on the lpSolveAPI package to solve the linear programming problem.
}
\examples{
\dontrun{
sparse_W <- dn_sparsifer(W)
}
}
