% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{entropy}
\alias{entropy}
\title{entropy}
\usage{
entropy(x, k)
}
\arguments{
\item{x}{data matrix used for calculating the entropy}

\item{k}{number for nearest neighbors used in entropy calculation}
}
\value{
a vector of entropy values
}
\description{
This subroutine estimates the entropy of a continuous random variable
}
\details{
\code{entropy} takes a continuous random variable and then estimates
entropy using the KSG estimator. 
It relies on the ANN package to query the kNN with KDTree algorithm.
}
