% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{estimate_turning_point}
\alias{estimate_turning_point}
\title{Calculate turning point for a linear or branched trajectory}
\usage{
estimate_turning_point(cds_subset, type = c("order", "value"),
  branch_point = 1, cores = 1)
}
\arguments{
\item{cds_subset}{a cds object after trajectory reconstruction}
}
\value{
a updated cds with a newly added column (turning_point) in pData indicates the inflection or branch time point.
to do: write a function to integrate BGP with Monocle
}
\description{
This function estimates the inflection point or the gene-wise branch point for each gene in a cell trajectory, without or with branch points respectively
}
