% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{knn_density}
\alias{knn_density}
\title{knn_density}
\usage{
knn_density(x, y, k)
}
\arguments{
\item{x}{1d vector of the data}

\item{y}{1d vector of querying points, positions used to estimate the density}

\item{k}{number of nearest neighbors used to estimate the 1d density, default to be 5}
}
\value{
a List where the element is the density estimate (name: density_estimate), 
the second one is the weight calculated based on density_estimate.
}
\description{
This subroutine calculates the 1d density of x at positions y
}
\details{
\code{knn_density} takes a vector of original data points x and a vector of querying points y
to calculate the density at each point y using the k-nearest neighbors.
}
