% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{knn_density_2d}
\alias{knn_density_2d}
\title{knn_density_2d}
\usage{
knn_density_2d(x, y, nGrids, k)
}
\arguments{
\item{x}{A vector for the values of the data on the first dimension}

\item{y}{A vector for the values of the data on the second dimension}

\item{nGrids}{A vector of two for the grid numbers on the first and second dimension}

\item{k}{number of nearest neighbors used to calculate the 2d density}
}
\value{
a numeric value for the d-dimensional unit ball for Euclidean norm
}
\description{
This subroutine calculates the density for a 2d space.
}
\details{
\code{knn_density_2d}
}
