% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lmi_multiple_run}
\alias{lmi_multiple_run}
\title{lmi_multiple_run}
\usage{
lmi_multiple_run(x, y, d, run_vec, umi)
}
\arguments{
\item{x}{one random variable from the time-series data}

\item{y}{another random variable from the time-series data}

\item{d}{Time lags used to estimate the RDI values}

\item{run_vec}{A integer vector encodes the information of the run id (run id start from 0)}

\item{uniformalize}{Whether or not you want to use ucmi to calculate rdi. Default to be false.}
}
\value{
a numeric value for the estimated mutual information between two variables (x, y) with a time lag d.
}
\description{
This subroutine calculates the lagged mutual information with multiple realization of the same processes.
}
\details{
\code{lmi_multiple_run} takes two random variables x and y, each has multiple realization of the same process and estimated their mutual information with a time lag d. 
using the KSG estimator. 
It relies on the ANN package to query the kNN with KDTree algorithm.
}
