% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lmi_single_run}
\alias{lmi_single_run}
\title{lmi_single_run}
\usage{
lmi_single_run(x, y, delay)
}
\arguments{
\item{x}{one random variable from the time-series data}

\item{y}{another random variable from the time-series data}

\item{delay}{Time lags used to estimate the RDI values}
}
\value{
a numeric value for the condition mutual information estimator between two variables (x, y) conditioned on variable z
}
\description{
This subroutine calculates the lagged mutual information
}
\details{
\code{cmi} takes two random variable x and y and estimated their mutual information conditioned on the third random variable z
using the KSG estimator. 
It relies on the ANN package to query the kNN with KDTree algorithm.
}
