% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mi}
\alias{mi}
\title{mi}
\usage{
mi(x, y, k, normalize)
}
\arguments{
\item{x}{one random variable from the time-series data}

\item{y}{another random variable from the time-series data}

\item{k}{number for nearest neighbors used in entropy calculation}

\item{normalize}{A logic flag to determine whether or not you want to normalize the MI value by the density of x.}
}
\value{
a estimated mutual information value between two variables (x, y)
}
\description{
This function estimates the mutual information of two random variables, x, y, based on their observed values
}
\details{
\code{mi} takes two random variables x and y to estimate the mutual information between them 
using the KSG estimator
It relies on the ANN package to query the kNN with KDTree algorithm.
}
