% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{new_arg}
\alias{new_arg}
\title{New argument}
\usage{
new_arg(
  id,
  aliases = NULL,
  actions = NULL,
  default = NULL,
  help = NULL,
  n = 0L
)
}
\arguments{
\item{id}{An integer to represent the sequence of argument}

\item{aliases}{One or more names to be associated with the argument}
}
\description{
Create a new argument class
}
