% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_causal_network}
\alias{plot_causal_network}
\title{Plot the causal network inferred by Scribe}
\usage{
plot_causal_network(graph, type = c("igraph", "hiearchy", "arcdiagram"),
  layout = NULL, ...)
}
\arguments{
\item{graph}{A igraph object describing inferred causal network}

\item{type}{A character describing which type of the plot will be used for making the network plot}

\item{layout}{A layout function which will only be used when we set type = 'igraph'.}
}
\value{
a plot object
}
\description{
This plot function uses state-of-art graph visualization tool (including plot, hiearchy plot and arc diagram) to visualize the causal network, identified by Scribe. 
See also the plot_hive_network for making beautiful Hive plots.
}
\examples{
\dontrun{
lung <- load_lung() 
gene_pairs_mat <- matrix(c('H19', 'Ccnd2', 'Ccnd2', 'Scnn1g'), ncol = 2)
plot_causal_network(lung, gene_pairs_mat)
}
}
\references{
Hive plot, hiearchy plot and arc diagram plot, etc.
}
