% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_comb_logic}
\alias{plot_comb_logic}
\title{Plot the combinatorial influence of two genes x, y to the target z.}
\usage{
plot_comb_logic(cds_subset, gene_pairs_target_mat, log = FALSE, d = 1,
  grid_num = 25, n_row = NULL, n_col = 1, normalized = TRUE,
  scales = "free", k = 5, verbose = FALSE)
}
\arguments{
\item{cds_subset}{CellDataSet for the experiment}

\item{n_row}{number of columns used to layout the faceted cluster panels}

\item{n_col}{number of columns used to layout the faceted cluster panels}

\item{k}{number of nearest points used in estimate the expected value}

\item{gene_pairs_mat}{A matrix where each row is the gene pair and the first column is the hypothetical source or regulator while the second column represents the hypothetical target. The name in this matrix should match the name in the gene_short_name column of the cds_subset object.}
}
\value{
a ggplot2 plot object
}
\description{
This plotting function tries to intuitively visualize the influence from genes x and y to the target z. 
Firstly, we divide the expression space for x and y based on grid_num and then we estimate the k-nearest neighbor for each of the
grid. We then use a Gaussian kernel to estimate the expected value for z. It is then displayed in two dimension with x and y
as two axis and the color represents the value of the expected of z. This function accepts a matrix where each row is the gene pair
and the target genes for this pair. The first column is the first hypothetical source or regulator, the second column represents 
the second hypothetical target while the third column represents the hypothetical target gene. The name in this matrix should match 
the name in the gene_short_name column of the cds_subset object.
}
\examples{
\dontrun{
lung <- load_lung() 
gene_pairs_mat <- matrix(c('H19', 'Ccnd2', 'Ccnd2', 'Scnn1g'), ncol = 2)
plot_comb_logic(lung, gene_pairs_mat)
}
}
