% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cross_map}
\alias{plot_cross_map}
\title{Plot the relationship between the predicted value (\hat{x}(t) | M_y) and the observed x(t).}
\usage{
plot_cross_map(cds_subset, gene_pairs_mat, tau = 1, E = 2,
  Type = c("Prediction", "Relationship"), verbose = FALSE,
  return_all = FALSE)
}
\arguments{
\item{cds_subset}{CellDataSet for the experiment}

\item{gene_pairs_mat}{A matrix where each row is the gene pair and the first column is the hypothetical source or regulator while the second column represents the hypothetical target. The name in this matrix should match the name in the gene_short_name column of the cds_subset object.}

\item{tau}{the lag to use for time delay embedding (same as in the ccm function from the rEDM package)}

\item{E}{the embedding dimensions to use for time delay embedding}

\item{Type}{}

\item{verbose}{whether or not to print the warning messages to the R console}
}
\value{
a ggplot2 plot object
}
\description{
This plotting function uses the simplex projection from the shadow manifold to estimate the source gene's value and then 
plot against the observe value. If there is a true causality from x to y, the predicted point should match up with 
the observed value pretty well. It also supports to plot the predicted value and the weighted target values based on 
the nearest neighbors used in the simplex projection. This two different plots are controlled by the type argument, 
where "Prediction" and "Relationship" corresponds to first and second plot respectively. This function accepts a matrix 
where each row is the gene pair and the first column is the hypothetical source or regulator while the second column 
represents the hypothetical target. The name in this matrix should match the name in the gene_short_name column of the 
cds_subset object.
}
\examples{
\dontrun{
lung <- load_lung() 
gene_pairs_mat <- matrix(c('H19', 'Ccnd2', 'Ccnd2', 'Scnn1g'), ncol = 2)
plot_cross_map(lung, gene_pairs_mat)
}
}
