% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gene_pairs}
\alias{plot_gene_pairs}
\title{Function to plot the trajectories of two genes}
\usage{
plot_gene_pairs(cds, gene_pairs, nrow = NULL, ncol = 1,
  panel_order = NULL)
}
\description{
Function to plot the trajectories of two genes
}
