% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gene_pairs_causality}
\alias{plot_gene_pairs_causality}
\title{Plot the heatmap for the expected value y(t) given x(t - d) and y(t - 1).}
\usage{
plot_gene_pairs_causality(cds_subset, gene_pairs_mat, log = FALSE, d = 1,
  k = 5, grid_num = 25, n_row = NULL, n_col = 1, scales = "free",
  verbose = FALSE)
}
\arguments{
\item{cds_subset}{CellDataSet for the experiment}

\item{gene_pairs_mat}{A matrix where each row is the gene pair and the first column is the hypothetical source or regulator while the second column represents the hypothetical target. The name in this matrix should match the name in the gene_short_name column of the cds_subset object.}

\item{log}{A logic argument to determine whether or not we should log the data before we perform density estimation.}

\item{d}{The time delay between the source gene and the target gene.}

\item{k}{The number of the nearest neighbors used in calculating the expectation value of y(t).}

\item{grid_num}{The number of grids used in estimating density.}

\item{n_row}{number of columns used to layout the faceted cluster panels}

\item{n_col}{number of columns used to layout the faceted cluster panels}

\item{scales}{}
}
\value{
a ggplot2 plot object
}
\description{
This plotting function tries to intuitively visualize the informatioin transfer from x(t - d) to y(t) given y(t)'s previous state y(t - 1).
Firstly, we divide the expression space for x(t - d) to y(t - 1) based on grid_num and then we estimate the k-nearest neighbor for each of the
grid. We then use a Gaussian kernel to estimate the expected value for y(t). It is then displayed in two dimension with x(t - d) and y(t - 1)
as two axis and the color represents the expected value of y(t) give x(t - d) and y(t - 1). This function accepts a matrix where each row is the gene pair and the 
first column is the hypothetical source or regulator while the second column represents the hypothetical target. The name in this matrix should 
match the name in the gene_short_name column of the cds_subset object.
}
\examples{
\dontrun{
lung <- load_lung() 
gene_pairs_mat <- matrix(c('H19', 'Ccnd2', 'Ccnd2', 'Scnn1g'), ncol = 2)
plot_gene_pair_causality(lung, gene_pairs_mat)
}
}
