% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gene_pairs_scatter}
\alias{plot_gene_pairs_scatter}
\title{Plot the scatterplot for gene pairs across different cells.}
\usage{
plot_gene_pairs_scatter(cds_subset, gene_pairs_mat, n_row = NULL, n_col = 1)
}
\arguments{
\item{cds_subset}{CellDataSet for the experiment}

\item{gene_pairs_mat}{A matrix where each row is the gene pair and the first column is the hypothetical source or regulator while the second column represents the hypothetical target. The name in this matrix should match the name in the gene_short_name column of the cds_subset object.}

\item{n_row}{number of columns used to layout the faceted cluster panels}

\item{n_col}{number of columns used to layout the faceted cluster panels}
}
\value{
a ggplot2 plot object
}
\description{
This plot function is the commomly used approach to visualize the relationship between gene pairs.
Please check other plots, including plot_gene_pairs_in_pseudotime, plot_gene_pairs_branched_pseudotime, plot_lagged_drevi, plot_cross_map, plot_gene_pairs_causality, plot_comb_logic,  
This function accepts a matrix where each row is the gene pair and the first column is the hypothetical source 
or regulator while the second column represents the hypothetical target. The name in this matrix should match 
the name in the gene_short_name column of the cds_subset object.
}
\examples{
\dontrun{
lung <- load_lung() 
gene_pairs_mat <- matrix(c('H19', 'Ccnd2', 'Ccnd2', 'Scnn1g'), ncol = 2)
plot_gene_pairs_scatter(lung, gene_pairs_mat)
}
}
