% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gene_pairs_scatter3D}
\alias{plot_gene_pairs_scatter3D}
\title{Plot the scatterplot formed by the space of x_{t - \mu}, y_t as well as y_{t - 1} for a single pair of genes.}
\usage{
plot_gene_pairs_scatter3D(cds_subset, gene_pair, d = 1, run_vec = NULL)
}
\arguments{
\item{cds_subset}{CellDataSet for the experiment}

\item{gene_pair}{A character vector consists of two gene short names while the first name is the hypothetical source and the second is the target.}

\item{d}{The time delay between the source and the target gene}

\item{run_vec}{A numeric vector of length equal to number of columns (cells) in the cds_subset, encoding the runs (or replicates) id for each experiment.}
}
\value{
a plotly object
}
\description{
This plot function is used to visualize the space formed by x_{t - \mu}, y_t as well as y_{t - 1} which is 
used to calculate the RDI score I(x_t -> Y_t) = CMI(x_{t - \mu}, y_t | y_{t - 1}). Please check other ploting
functions (plot_gene_pairs_in_pseudotime, plot_gene_pairs_branched_pseudotime, plot_lagged_drevi, plot_cross_map, plot_gene_pairs_causality, plot_comb_logic) for more intuitive 
visualization.
}
\examples{
\dontrun{
lung <- load_lung() 
gene_pairs_mat <- matrix(c('H19', 'Ccnd2', 'Ccnd2', 'Scnn1g'), ncol = 2)
plot_gene_pairs_scatter3D(lung, gene_pairs_mat[1, ], d = 5)
}
}
