% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_hive_network}
\alias{plot_hive_network}
\title{Make a hive plot}
\usage{
plot_hive_network(gene_names, casuality_mat_list, cds_list, group_names,
  data_group_color = c("#E41A1C", "#377EB8", "#4DAF4A", "#984EA3", "#FF7F00",
  "#6A3D9A", "#A65628", "#F781BF"), top_edge_num = 100,
  axLab.gpar = grid::gpar(col = "black", fontsize = 6, lwd = 0.5),
  bkgnd = "white", return_hive_plot_data = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{gene_names}{A vector of gene names whose length should match up with the dimensions of the matrix in the casuality_mat_list.}

\item{casuality_mat_list}{A list of square matrix of the casuality score (either RDI or cRDI, etc.).}

\item{cds_list}{A list of cell dataset objects correspond to the each of the casuality score matrix.}

\item{group_names}{A vector of group names correspond to the data in the list, which will be used to label the axis on the Hive plot.}

\item{top_edge_num}{Number of top edges you would like to select from each dataset to visualize. Default to be 100.}

\item{axLab.gpar}{(Same argument in plotHive function) A list of name - value pairs acceptable to gpar. These control the label and arrow displays.}

\item{bkgnd}{(Same argument in plotHive function) Any valid color specification. Used for the background color for plotHive.}

\item{return_hive_plot_data}{A logic flag to determine whether or not you would like to return the Hive plot object}

\item{verbose}{A logic flag to determine whether or not you would like to see detailed running information}

\item{...}{Other arguments passed into the plotHive function.}

\item{group_color}{A vector of color correspond to the data in the list, which will be used to color the edges for each group of data.}
}
\value{
a hiveplot object
}
\description{
This function creates a HiveR plot with the genes names, list of causality score matrix for each dataset (cell types), each cell dataset and 
name of each cell datasets and the color to visualize each dataset. It is based on the HiveR package.
}
\examples{
\dontrun{
gene_list <- c(positive_score_genes, negtive_score_genes)
cell_type_list <- c("Erythrocyte", "Megakaryocyte", "Monocyte", "Neutrophil", 'Basophil / Eosinophil', 'Dendritic cell')
cell_type_color <- c("#E41A1C", "#377EB8", "#4DAF4A", "#984EA3", "#FF7F00", "#6A3D9A", "#A65628", "#F781BF")

rdi_list <- list(Ery_list$max_rdi_value, MK_list$max_rdi_value, Mono_list$max_rdi_value, Neu_list$max_rdi_value, BE_list$max_rdi_value, DC_list$max_rdi_value)
cds_list <- list(Ery_cds, MK_cds, Mono_cds, Neu_cds, BE_cds, DC_cds)
plot_hive_network(gene_list, rdi_list, cds_list, cell_type_list, top_edge_num = 25)
}
}
