% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_lineage_expression}
\alias{plot_lineage_expression}
\title{Visualize gene expression level over lineage tree}
\usage{
plot_lineage_expression(cds, gene_names, ddata, lineage_tree, lineage_coord,
  root_cell, lineage_depth = NULL, expression_threshold = NULL,
  show_lineage_pair = FALSE, label_x = 0.45, label_y = 1,
  return_all = FALSE, ...)
}
\arguments{
\item{cds}{A cell dataset object. This cds should include genes from gene_names variable (by row
names) and cells from the lineage_tree variable.}

\item{gene_names}{The genes you would like to visualize the lineage expression.}

\item{ddata}{A matrix containing information for the coordinates of all segments on the lineage tree.}

\item{lineage_tree}{An igraph object containing the lineage relationship for all the cells in the cds object.}

\item{lineage_coord}{A matrix containing the coordinates for the cells from the cds object.}

\item{root_cell}{The root cells from where the lineage will be drawn.}

\item{lineage_depth}{The maximum time point for the lineage tree you would like to draw. Default to be NULL (no limitation on the lineage depth to be drawn).}

\item{expression_threshold}{The threshold cutoff to determine whether or not a gene is expressed. Default to be 2000 (A reasonable cutoff based on the background fluorescence signal intensity on the c. elegans fluorescence data.)}

\item{show_lineage_pair}{A logic argument to determine whether or not you would like to show the two immediate branch from the root_cell as a lineage pair. The lineage pair will be shown on two rows while the second row is flip for the purpose of direct comparison.}

\item{label_x}{Parameters passed to the plot_grid function from cowplot package.}

\item{label_y}{Parameters passed to the plot_grid function from cowplot package.}

\item{return_all}{A logic argument to determine whether or not to return all the metadata relevant to creating the lineage expression plot.}

\item{...}{Other arguments. Currently not used.}

\item{verbose}{A logic argument to determine whether or not to print all running informations.}
}
\value{
a ggplot object
}
\description{
This function takes gene expression values (stored in the CDS object) for one or a few genes (gene_names) and visualize the 
expression values or status of expression (either expressed or not expressed) over a lineage tree (lineage_tree), starting 
from the defined root cell (root_cell) all the way to the end of the lineage, where the coordinates for tree is stored in 
the ddata matrix.
}
\examples{
\dontrun{

plot_lineage_expression(c_elegans_cds, max_time = 250, matrix(c('hnd.1', 'unc.120'), nrow = 1))
}
}
