% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_shadow_manifold}
\alias{plot_shadow_manifold}
\title{Plot the shadow mainfold from the lagged coordinates from a single variable.}
\usage{
plot_shadow_manifold(cds_subset, gene_name, d = 1)
}
\arguments{
\item{cds_subset}{CellDataSet for the experiment}

\item{gene_name}{The name for gene used to reconstruct the shadow manifold. The gene name should appear in the gene_short_name column of the cds_subset object.}

\item{d}{The time delay between the source and target gene.}
}
\value{
a ggplot2 plot object
}
\description{
This plot function uses time lagged coordinates of a single variable (gene) x, (x_{t}, x_{t - d}, x_{t - 2d}) to reconstruct a shadow manifold. In CCM (cross-convergence mapping), this manifold can 
be used for detecting causality.
}
\examples{
\dontrun{
x_1 <- all_cell_simulation[1, , 1]
x_2 <- all_cell_simulation[2, , 1]
x_3 <- all_cell_simulation[3, , 1]

plot_ly(type = 'scatter3d', x = x_1, y = x_2, z = x_3, mode = 'markers', color = 1:length(x_1))
lung <- load_lung() 
gene_pairs_mat <- matrix(c('H19', 'Ccnd2', 'Ccnd2', 'Scnn1g'), ncol = 2)
plot_shadow_manifold(lung, gene_pairs_mat[1])
}
}
\references{
Sugihara, George, et al. "Detecting causality in complex ecosystems." science 338.6106 (2012): 496-500.
}
