% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_spatial_causality}
\alias{plot_spatial_causality}
\title{Visualize the spatial causality}
\usage{
plot_spatial_causality(full_cds, max_time = NULL, gene_pairs,
  uniformalize = TRUE, return_all = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{full_cds}{The cds includes all the data points. This cds needs to include information about time, the x, y, z coordinates for the cell.}

\item{max_time}{Maximum time point considered for the cell dataset.}

\item{gene_pairs}{A matrix containing the gene-pair interactions you want to visualize across spatial locations. Need to be two columns. 
The first column is the source while the second column is the target.}

\item{return_all}{A logic argument to determine whether or not to return both the plotly result as well as the data.frame used to make this plot.}

\item{verbose}{A logic argument to determine whether or not to print all running informations.}

\item{...}{Other arguments. Currently not used.}
}
\value{
a plotly object
}
\description{
This function creates a HiveR plot with the genes names, list of causality score matrix for each dataset (cell types), each cell dataset and 
name of each cell datasets and the color to visualize each dataset. It is based on the HiveR package.
}
\examples{
\dontrun{
plot_sp_causality(c_elegans_cds, max_time = 250, matrix(c('hnd.1', 'unc.120'), nrow = 1))
}
}
