% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_taylor_diagram}
\alias{plot_taylor_diagram}
\title{Plot the taylor diagram for the RDI, mutual information values.}
\usage{
plot_taylor_diagram(ref, RDI, MI, Reference, add = FALSE, col = "red",
  pch = 19, pos.cor = TRUE, xlab = "Entropy", ylab = "MI",
  main = "RDI Taylor Diagram", show.gamma = TRUE, ngamma = 3,
  gamma.col = 8, sd.arcs = 0, ref.sd = FALSE, sd.method = "sample",
  grad.corr.lines = c(0.2, 0.4, 0.6, 0.8, 0.9), pcex = 1, cex.axis = 1,
  normalize = FALSE, mar = c(5, 4, 6, 6), text = NULL, ...)
}
\arguments{
\item{cds_subset}{CellDataSet for the experiment}

\item{gene_pairs_mat}{A matrix where each row is the gene pair and the first column is the hypothetical source or regulator while the second column represents the hypothetical target. The name in this matrix should match the name in the gene_short_name column of the cds_subset object.}

\item{n_row}{number of columns used to layout the faceted cluster panels}

\item{n_col}{number of columns used to layout the faceted cluster panels}
}
\value{
a ggplot2 plot object
}
\description{
This plot function is the commomly used approach to visualize the relationship between a gene pair.
Please check other plots for plot_lagged_drevi, plot_cross_map, plot_gene_pairs_causality, plot_comb_logic,  
This function accepts a matrix where each row is the gene pair and the first column is the hypothetical source 
or regulator while the second column represents the hypothetical target. The name in this matrix should match 
the name in the gene_short_name column of the cds_subset object.
}
\examples{
\dontrun{
lung <- load_lung() 
gene_pairs_mat <- matrix(c('H19', 'Ccnd2', 'Ccnd2', 'Scnn1g'), ncol = 2)
plot_taylor_diagram(cds_exprs[1, ], seq(0, 1, length.out = 100), MI = seq(2, 3, length.out = 100), Entropy = seq(4, 5, length.out = 1))
}
}
