% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_time_delay}
\alias{plot_time_delay}
\title{Plot a heatmap to demonstrate the hypothetical time delay between any two genes}
\usage{
plot_time_delay(cds_subset, use_gene_short_name = TRUE, cluster_row = TRUE,
  cluster_cols = TRUE)
}
\arguments{
\item{cds_subset}{CellDataSet for the experiment}

\item{gene_pairs_mat}{A matrix where each row is the gene pair and the first column is the hypothetical source or regulator while 
the second column represents the hypothetical target. The name in this matrix should match the name in the gene_short_name column 
of the cds_subset object.}

\item{n_row}{number of columns used to layout the faceted cluster panels}

\item{n_col}{number of columns used to layout the faceted cluster panels}
}
\value{
a ggplot2 plot object
}
\description{
This plot function uses the estimated turning_point for each gene in the cds_subset to estimate the time delay between any gene pairs. 
It then plots the time delay using a heatmap.
}
\examples{
\dontrun{
lung <- load_lung() 
gene_pairs_mat <- matrix(c('H19', 'Ccnd2', 'Ccnd2', 'Scnn1g'), ncol = 2)
plot_time_delay(lung, gene_pairs_mat)
}
}
