% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdi_many_runs}
\alias{rdi_many_runs}
\title{rdi_many_runs}
\usage{
rdi_many_runs(x, y)
}
\arguments{
\item{x}{a random variable with multiple run of the same process}

\item{y}{another random variable with multiple run of another process}
}
\value{
a numeric vector storing the DI from two multiple run variables
}
\description{
This function simulates the DIRECTED mutual information from X to Y when you have multiple run of the processes
}
\details{
\code{rdi_many_runs} takes two random variables (with the same multiple realization of the two processes) and estimate the direct information between them. 
using the formula: 0.5 * d * log(pi) - log(gamma(0.5 * d + 1))
It's implimented in C++, providing a (small) increase in speed over the R equivalent.
}
