% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdi_multiple_runs_conditioned}
\alias{rdi_multiple_runs_conditioned}
\title{This function simulates the CONDITIONED DIRECTED mutual information from X to Y CONDITIONED ON Z when you have a SINGLE run of the processes}
\usage{
rdi_multiple_runs_conditioned(x, y, z, z_delays, d, run_vec, uniformalize)
}
\arguments{
\item{x}{one random variable from the time-series data}

\item{y}{another random variable from the time-series data}

\item{z}{z is a dataframe or matrix consisting of the data for different variables}

\item{z_delays}{z_delay is also a dataframe or matrix consisting of the delays to be applied to different variables}

\item{d}{delay in the formula I(x-->y)=I(x_{t-d};y_t|y_{t-1}) default: d=1}

\item{uniformalize}{Whether or not you want to use ucmi to calculate rdi. Default to be false.}
}
\value{
a matrix for the condition mutual information estimators between all pairwise variables (x, y) in the data matrix x, y
}
\description{
This subroutine calculates the volume of a d-dimensional unit ball for Euclidean norm
}
\details{
\code{rdi_single_run_conditioned} takes two random variables x and y as well as the parameter n to calculate the restricted direct information conditioned on variable z.
}
