% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdi_single_run}
\alias{rdi_single_run}
\title{rdi_single_run}
\usage{
rdi_single_run(x, y, d)
}
\arguments{
\item{x}{one random variable from the time-series data}

\item{y}{another random variable from the time-series data}

\item{d}{delay in the formula I(x-->y)=I(x_{t-d};y_t|y_{t-1}) default: d=1}
}
\value{
a matrix for the condition mutual information estimators between all pairwise variables (x, y) in the data matrix x, y
}
\description{
This function simulates the RESTRICTED DIRECTED mutual information from X to Y when you have a SINGLE run of the processes
}
\details{
\code{rdi_single_run} takes two random variables x and y as well as a time delay d to estimate the restricted direct infomation between them.
}
