% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-args.R
\docType{class}
\name{scribeArg-class}
\alias{scribeArg-class}
\alias{scribeArg}
\title{{scribe} argument}
\description{
ReferenceClass object for managing arguments
}
\section{Fields}{

\describe{
\item{\code{aliases}}{A character vector to denote the argument's name}

\item{\code{action}}{An action for resolving the argument}

\item{\code{options}}{A named list of options (see details)}

\item{\code{convert}}{Passed to the \code{to} argument in \code{\link[=value_convert]{value_convert()}}}

\item{\code{default}}{A default value}

\item{\code{info}}{Additional information about the argument when printed}

\item{\code{n}}{The length of the values}

\item{\code{id}}{An integer id (used when stored within \link{scribeCommandArgs})}
}}

\section{Methods}{

\describe{
\item{\code{get_action()}}{Retrieve action}

\item{\code{get_aliases()}}{Retrieve aliases}

\item{\code{get_default()}}{Retrieve the default value}

\item{\code{get_help()}}{Retrieve help information as a character vector}

\item{\code{get_name(clean = TRUE)}}{Retrieve names

    \describe{
      \item{\code{clean}}{When \code{TRUE} removes \code{-}s from text}
    }}

\item{\code{help()}}{Print out formatted help information}

\item{\code{initialize(
  aliases = NULL,
  id = NULL,
  action = NULL,
  options = NULL,
  convert = NULL,
  default = NULL,
  info = NULL,
  n = NULL
)}}{    Initialize the \code{scribeArg} object

    See \strong{fields} for parameter information.
    }

\item{\code{parse_value(command_arg)}}{Parse argument value

    This method will likely not be called directly by the user.  Instead, this
    method is called within [scribeCommandArg].

    \describe{
      \item{\code{command_arg}}{A \code{scribeCommandArg}} object
    }}
}}

\section{Options}{

Available options include:
\describe{
\item{\code{action="list"}}{
\describe{
\item{\code{choices}}{
An explicit set of values that argument must be.  If the value parsed
is not one of these, an error will occur.
}
}
}
\item{\code{action="flag"}}{
\describe{
\item{\code{no}}{
When \code{TRUE} included appends \code{--no} to aliases to invert results\cr
\strong{Example:}\cr
With the arg \code{new_arg("--test", options = list(no = TRUE))}, passing
command args \code{--test} would set this to \code{TRUE} and \code{--no-test}
explicitly set to \code{FALSE}.
}
}
}
}
}

