% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-command-args.R
\docType{class}
\name{scribeCommandArgs-class}
\alias{scribeCommandArgs-class}
\alias{scribeCommandArgs}
\title{{scribe} command arguments}
\description{
Reference class object for managing command line arguments.
}
\section{Fields}{

\describe{
\item{\code{input}}{\verb{[character]}\cr A character vector of command line arguments. See also
\code{\link[=command_args]{command_args()}}}

\item{\code{working}}{\verb{[character]}\cr A copy of \code{input}.  Note: this is used to track parsing
progress and is not meant to be accessed directly.}

\item{\code{values}}{\verb{[list]}\cr A named \code{list} of values.  Empty on initialization and
populated during argument resolving.}

\item{\code{argList}}{\verb{[list]}\cr a List of \link{scribeArg}s}

\item{\code{nArgs}}{\verb{[interger]}\cr A current counter for the number of \link{scribeArg}s included}

\item{\code{resolved}}{\verb{[logical]}\cr A \code{logical} value indicated if the \code{resolve()} method has
been successfully executed.}

\item{\code{description}}{\verb{[character]}\cr Additional help information}

\item{\code{included}}{\verb{[character]}\cr Default \link{scribeArg}s to include}

\item{\code{examples}}{\verb{[character]}\cr Examples to print with help}

\item{\code{comments}}{\verb{[character]}\cr Comments printed with}
}}

\section{Methods}{

\describe{
\item{\code{add_argument(
  ...,
  action = arg_actions(),
  options = NULL,
  convert = default_convert,
  default = NULL,
  n = NA_integer_,
  info = NULL,
  id = NULL
)}}{Add a \link{scribeArg} to \code{argList}

    \describe{
      \item{\code{...}}{Aliases}
      \item{(other)}{See \code{\link[=new_arg]{new_arg()}} for details}
    }}

\item{\code{add_description(..., sep = "")}}{Add a value to \code{description}

    \describe{
      \item{\code{...}}{Information to paste into the description}
      \item{\code{sep}}{\code{character} separate for \code{...}}
    }}

\item{\code{add_example(x, comment = "", prefix = "$ ")}}{Add a value to \code{examples}

    \describe{
      \item{\code{x}}{A code example as a \code{character}}
      \item{\code{comment}}{An optional comment to append}
      \item{\code{prefix}}{An optional prefix for the example}
    }}

\item{\code{append_arg(arg)}}{Append a new \link{scribeArg} value to \code{argList}

    \describe{
      \item{\code{arg}}{The \link{scribeArg} to include}
    }}

\item{\code{arg_counter()}}{Increase the value of \code{nArgs} by \code{1L}}

\item{\code{get_args(included = TRUE)}}{Retrieve \code{argList}}

\item{\code{get_description()}}{Retrieve \code{description}}

\item{\code{get_examples()}}{Retrieve \code{examples}}

\item{\code{get_input()}}{Retrieve \code{input}}

\item{\code{get_n_args()}}{Retrieve the current number of \link{scribeArg}s}

\item{\code{get_values()}}{Retrieve \code{values}}

\item{\code{get_working()}}{Retrieve \code{working}}

\item{\code{help()}}{Print the help information}

\item{\code{initialize(input = "", include = c("help", "version", NA_character_))}}{    Initialize the \link{scribeCommandArgs} object

    See \strong{fields} for parameter information.
    }

\item{\code{parse()}}{Return the parsed values of from each \code{argList}}

\item{\code{remove_working(i)}}{Retrieve \code{working}

    \describe{
      \item{\code{i}}{Index value of \code{working} to remove}
    }}

\item{\code{resolve()}}{Resolve the values of each \link{scribeArg} in \code{argList}}

\item{\code{set_description(..., sep = "")}}{Set the value of \code{description}

    \describe{
      \item{\code{...}}{Information to paste into the description}
      \item{\code{sep}}{\code{character} separate for \code{...}}
    }}

\item{\code{set_example(x = character(), comment = "", prefix = "$ ")}}{Set the value of \code{examples}

    \describe{
      \item{\code{x}}{A code example as a \code{character}}
      \item{\code{comment}}{An optional comment to append}
      \item{\code{prefix}}{An optional prefix for the example}
    }}

\item{\code{set_input(value)}}{Set \code{input}

    \describe{
      \item{\code{value}}{Value to set}
    }}

\item{\code{set_values(i = TRUE, value)}}{Set \code{values}

    \describe{
      \item{\code{i}}{Index value of \code{working} to set}
      \item{\code{value}}{The value to set}
    }}

\item{\code{version()}}{Print the \link{scribe-package} version}

\item{\code{write_usage()}}{Write usage information to a \code{character} vector}
}}

