% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smooth_gene}
\alias{smooth_gene}
\title{entropy}
\usage{
smooth_gene(expr_data, window_size = 40L)
}
\arguments{
\item{R_x}{number of dimension}

\item{R_k}{number of dimension}
}
\value{
a updated matrix with gene expression smoothed with window size equal to window_size
}
\description{
This subroutine calculates the volume of a d-dimensional unit ball for Euclidean norm
}
\details{
\code{entropy} takes a integer of dimensions and then calculate the olume of a d-dimensional unit ball for Euclidean norm
using the formula: 0.5 * d * log(pi) - log(gamma(0.5 * d + 1))
It's implimented in C++, providing a (small) increase in speed over the R equivalent.
}
