% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{umi}
\alias{umi}
\title{umi}
\usage{
umi(x, y, k, method, k_density, bw)
}
\arguments{
\item{x}{one random variable from the time-series data}

\item{y}{another random variable from the time-series data}

\item{k}{number for nearest neighbors used in entropy calculation}

\item{method}{Which 2D density estimator you would like to use. 1 is kde estimator and 2 is knn based estimator.}

\item{k_density}{The number of k nearest neighbors you would like to use when calculating the density 
(only applicable when method == 2 or using knn based density estimation).}

\item{bw}{Bindwidth used for the kernel density estimator. Currently it is not used. The bindwidth in the kde function is automatically estimated.}
}
\value{
A estimated uniform mutual information value between two variables (x, y) where the distribution for the x is replaced by a uniform distribution.
}
\description{
This subroutine calculates the uniformed  mutual information where 
the distribution for x is replaced by a uniform distribution.
}
\details{
\code{umi} takes two random variable x and y and estimated their mutual using the KSG estimator while x is replaced by a uniform distribution. 
It relies on a C++ implmentation of kde estimator (https://github.com/timnugent/kernel-density) and the ANN package to query the kNN with KDTree algorithm.
}
