% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vd}
\alias{vd}
\title{vd}
\usage{
vd(d)
}
\arguments{
\item{d}{number of dimension}
}
\value{
a numeric value for the d-dimensional unit ball for Euclidean norm
}
\description{
This subroutine calculates the volume of a d-dimensional unit ball for Euclidean norm
}
\details{
\code{vd} takes a integer of dimensions and then calculate the volume of a d-dimensional unit ball for Euclidean norm
using the formula: 0.5 * d * log(pi) - log(gamma(0.5 * d + 1))
It's implimented in C++, providing a (small) increase in speed over the R equivalent.
}
