% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wired.R
\name{wired}
\alias{wired}
\title{Causal network inference limitted to the interaction between informative transcription factors and those to other informative target genes.}
\usage{
wired(cds, TF, informative_genes, delays = c(1, 10, 15),
  include_conditioning = FALSE, smoothing = FALSE, cluster_TFs = FALSE,
  cluster_targets = FALSE, cluster_TFs_num = NULL,
  cluster_targets_num = NULL, cluster_num_method = c(NULL, "mcclust",
  "pamk"), hclust_method = "ward.D2", norm_method = c("log", "vstExprs"),
  scale = FALSE, scale_max = 3, scale_min = -3, ...)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{TF}{A vector of the gene short names for all the transcription factors.}

\item{informative_genes}{A vector of the informative genes used for network inference, which is identified through BEAM or other methods. It should include some TFs and normally other target genes.}

\item{delays}{A vector of time delays used during information transfer estimation between genes}

\item{include_conditioning}{A logic argument to determine whether or not we should also calculate the conditioning RDI values}

\item{smoothing}{A logic argument to determine whether or not we should also smooth the data using loess before we calculate the RDI values}

\item{cluster_TFs}{A logic value to determine whether or not we should cluster the TFs' expression (for either the group of TFs or targets) before estimating the RDI/cRDI value.}

\item{cluster_targets}{A logic value to determine whether or not we should cluster the target genes' expression (for either the group of TFs or targets) before estimating the RDI/cRDI value.}

\item{cluster_TFs_num}{Number of TF clusters you will use for estimating the information transfer.}

\item{cluster_targets_num}{Number of target clusters you will use for estimating the information transfer.}

\item{cluster_num_method}{Which method you will use to determine the proper number of cluster for either the TFs or the targets. 
Options include NULL (default, don't use any method), mcclust (model-based optimal number of clusters: TF) and finally pamk (number of clusters estimated by optimum average silhouette width).}

\item{hclust_method}{Which cluster method in hclust you will use. Default to be ward.D2.}

\item{norm_method}{Determines how to transform expression values prior to estimate information transfer}

\item{scale}{Determine whether or not to scale data prior to estimate information transfer}

\item{scale_max}{The maximum value (in standard deviations) to show in the heatmap. Values larger than this are set to the max.}

\item{scale_min}{The minimum value (in standard deviations) to show in the heatmap. Values smaller than this are set to the min.}
}
\value{
a matrix storing the RDI values for interactions from TFs to the targets as well as between transcription factors.
}
\description{
This function implements a procedure to build a causal network based on some prior information of the transcription factors (TFs). It calculates the information 
transfer from TFs to the targets as well as the information transfer between transcription factors while avoiding the calculation of information transfer between 
the targets and from targets to the TFs. This function accepts a vector for the gene short names for all the transcription factors and another vector for the 
informative genes selected through BEAM or methods.
}
\examples{
\dontrun{
lung <- load_lung() 
TF <- c('H19', 'Ccnd2', 'Ccnd2', 'Scnn1g'); TF_id <- row.names(subset(fData(lung), gene_short_name \%in\% TF))
informative_genes_id <- row.names(lung)
wired(lung, TF_id, informative_genes_id)
}
}
