% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{calculate_rdi}
\alias{calculate_rdi}
\title{calculate_rdi}
\usage{
calculate_rdi(genes_data, delays, super_graph = NULL, turning_points = NULL,
  method = 1)
}
\arguments{
\item{delays}{Time lags used to estimate the RDI values}

\item{method}{either 1, 2 represents either using (RDI: restricted direction information) or LMI (lagged mutual information)}

\item{Time}{/ Pseudo-time series expression data where each row is a sample, each column is a feature}
}
\value{
a list with four components: matrix for RDI (dimension is number of genes X length of delays times number of genes), vector of delays, max_rdi_value and max_rdi_delays
}
\description{
This subroutine calculates RDI of Pseudo-time series expression data
}
\details{
\code{calculate_rdi} takes a time / Pseudo-time series expression data as well as the time lags and then calculate the restricted direct information between each pair of genes under different delays.
}
