% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{calculate_rdi_multiple_run}
\alias{calculate_rdi_multiple_run}
\title{calculate_rdi_multiple_run}
\usage{
calculate_rdi_multiple_run(cds_subset, run_vec = NULL, delays,
  super_graph = NULL, method = 1, uniformalize = FALSE, log = TRUE,
  pseudo_cnt = 1)
}
\arguments{
\item{cds_subset}{A cds_subset which has been ordered with Monocle 2}

\item{run_vec}{A vector describes which run (lineage) does the current cell come from. It has the some length as the row length of genes_data.}

\item{delays}{Time lags used to estimate the RDI values (be ready to update this as a matrix or a vector with the same length as the run_vec once we have a way to directly calculate the proper time lag need to be used)}

\item{super_graph}{A graph including all possible interactions used for performing the causality inference. When it is NULL, 
all possible pairs of interactions for the genes_data are used, otherwise only the interactions specified in the graph are used.
Note that the super_graph only accept integer matrix for now (each integer corresponds to a particular gene in the genes_data).}

\item{method}{either 1, 2 represents either using (RDI: restricted direction information) or LMI (lagged mutual information)}

\item{uniformalize}{Whether or not you want to use ucmi to calculate rdi. Default to be false.}
}
\value{
a list with four components: matrix for RDI (dimension is number of genes X length of delays times number of genes), vector of delays, max_rdi_value and max_rdi_delays
}
\description{
This subroutine estimates the RDI value for all gene-pair combination from the genes_data (or that specified in the super_graph) 
in the pseudotime/time series data. If multiple directly related experiments are conducted, we can pass the information of each 
experiment by the run_vec argument and this function will concatenate those experiments based on run_vec.
}
\details{
\code{calculate_rdi_multiple_run} takes a time / Pseudo-time series expression data as well as the time lags and then calculate the restricted direct information between each pair of genes under different delays.
}
\examples{
\dontrun{
lung <- load_lung() 
lung_dup <- buildBranchCellDataSet(lung)
run_vec <- pData(lung_dup)$Branch
Branch_uniq <- unique(run_vec)
run_vec[run_vec == Branch_uniq[1]] <- 0
run_vec[run_vec == Branch_uniq[2]] <- 1
lung_res <- calculate_rdi_multiple_run(lung, run_vec = run_vec)
}
}
