% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-command-args.R
\name{command_args}
\alias{command_args}
\title{Command args}
\usage{
command_args(
  x = commandArgs(trailingOnly = TRUE),
  include = c("help", "version")
)
}
\arguments{
\item{x}{Command args; see \code{\link[base:commandArgs]{base::commandArgs()}} for default}

\item{include}{Special default arguments to included}
}
\value{
A \code{scribeCommandArgs} Reference object
}
\description{
Command args
}
\concept{scribe}
