% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{command_args}
\alias{command_args}
\title{Command args}
\usage{
command_args(x = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{x}{Command args; see \code{\link[base:commandArgs]{base::commandArgs()}} for default}
}
\value{
A \code{scribeCommandArgs} Reference object
}
\description{
Command args
}
