% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{di_single_run_conditioned}
\alias{di_single_run_conditioned}
\title{di_single_run_conditioned}
\usage{
di_single_run_conditioned(x, y, z, n)
}
\arguments{
\item{x}{one random variable from the time-series data}

\item{y}{another random variable from the time-series data}

\item{z}{z is a dataframe (or matrix) containing the data of other processes upon the past of which the mi is conditioned}

\item{n}{Parameter n determines the the number of previous time samples upon which the mi is conditioned}
}
\value{
a matrix for the condition mutual information estimators between all pairwise variables (x, y) in the data matrix x, y
}
\description{
This function simulates the CONDITIONED DIRECTED mutual information from X to Y when you have a SINGLE run of the processes
}
\details{
\code{di_single_run_conditioned} takes two random variables x and y as well as the parameter n to calculate the direct information conditioned on variable z.
}
