% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lmi_single_run}
\alias{lmi_single_run}
\title{lmi_single_run}
\usage{
lmi_single_run(x, y, delay, umi)
}
\arguments{
\item{x}{one random variable from the time-series data}

\item{y}{another random variable from the time-series data}

\item{delay}{Time lags used to estimate the RDI values}

\item{uniformalize}{Whether or not you want to use ucmi to calculate rdi. Default to be false.}
}
\value{
a numeric value for the mutual information estimator between two variables (x, y) with a time lag d.
}
\description{
This subroutine calculates the lagged mutual information
}
\details{
\code{lmi_single_run} takes two random variable x and y and estimated their mutual information with a time lag d. 
using the KSG estimator. 
It relies on the ANN package to query the kNN with KDTree algorithm.
}
