% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mi}
\alias{mi}
\title{mi}
\usage{
mi(x, y, k)
}
\arguments{
\item{x}{one random variable from the time-series data}

\item{y}{another random variable from the time-series data}

\item{k}{number for nearest neighbors used in entropy calculation}
}
\value{
a numeric value for the mutual information estimator between two variables (x, y)
}
\description{
This function estimates the mutual information of two random variables based on their observed values
}
\details{
\code{mi} takes two random variables x and y to estimate the mutual information between them 
using the KSG estimator
It relies on the ANN package to query the kNN with KDTree algorithm.
}
