% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-args.R
\name{new_arg}
\alias{new_arg}
\title{New Arg}
\usage{
new_arg(
  aliases = "",
  action = arg_actions(),
  convert = default_convert,
  options = list(),
  default = NULL,
  info = character(),
  n = NA_integer_,
  id = NA_integer_
)
}
\arguments{
\item{aliases}{A \code{character} vector to denote the argument's name}

\item{action}{An action for resolving the argument}

\item{convert}{Passed to the \code{to} argument in \code{\link[=value_convert]{value_convert()}}}

\item{options}{A named list of options (see \link{scribeArg})}

\item{default}{A default value}

\item{info}{Additional information about the argument when printed}

\item{n}{The length of the values}

\item{id}{An integer id (used when stored within \link{scribeCommandArgs})}
}
\value{
A \link{scribeArg} object
}
\description{
Make a new \link{scribeArg} object
}
