% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-args.R
\name{new_arg}
\alias{new_arg}
\title{New command argument}
\usage{
new_arg(
  aliases = "",
  action = arg_actions(),
  convert = default_convert,
  options = list(),
  default = NULL,
  info = NULL,
  n = NA_integer_
)
}
\arguments{
\item{aliases, action, convert, options, default, info, n, id}{See \verb{$initialize()}
in \link{scribeArg}.}
}
\value{
A \link{scribeArg} object
}
\description{
Make a new \link{scribeArg} object
}
