% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{new_arg}
\alias{new_arg}
\title{New command argument}
\usage{
new_arg(
  aliases = "",
  action = arg_actions(),
  default = NULL,
  convert = default_convert,
  n = NA_integer_,
  info = NULL,
  options = list()
)
}
\arguments{
\item{aliases, action, convert, options, default, info, n}{See \verb{$initialize()}
in \link{scribeArg}.}
}
\value{
A \link{scribeArg} object
}
\description{
Make a new \link{scribeArg} object
}
\examples{
new_arg()
new_arg("values", action = "dots")
new_arg(c("-f", "--force"), action = "flag")
}
\seealso{
Other scribe: 
\code{\link{command_args}()},
\code{\link{scribeArg-class}},
\code{\link{scribeCommandArgs-class}}
}
\concept{scribe}
