% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gene_pairs_in_pseudotime}
\alias{plot_gene_pairs_in_pseudotime}
\title{Plot the expression dynamics for gene pairs in pseduotime with one separate kinetic curve for each gene.}
\usage{
plot_gene_pairs_in_pseudotime(cds_subset, gene_pairs_mat, min_expr = NULL,
  cell_size = 1.5, n_row = NULL, n_col = 1, color_by = "State",
  trend_formula = "~ sm.ns(Pseudotime, df=3)", label_by_short_name = TRUE,
  relative_expr = TRUE, vertical_jitter = NULL, horizontal_jitter = NULL,
  fitting_type = c("loess", "spline"), log = TRUE, turning_point = FALSE,
  data_points = FALSE)
}
\arguments{
\item{cds_subset}{CellDataSet for the experiment}

\item{gene_pairs_mat}{A matrix where each row is the gene pair and the first column is the hypothetical source or regulator while the second column represents the hypothetical target. The name in this matrix should match the name in the gene_short_name column of the cds_subset object.}

\item{min_expr}{the minimum (untransformed) expression level to use in plotted the genes.}

\item{cell_size}{the size (in points) of each cell used in the plot}

\item{n_row}{number of columns used to layout the faceted cluster panels}

\item{n_col}{number of columns used to layout the faceted cluster panels}

\item{color_by}{the cell attribute (e.g. the column of pData(cds_subset)) to be used to color each cell}

\item{trend_formula}{the model formula to be used for fitting the expression trend over pseudotime}

\item{label_by_short_name}{label figure panels by gene_short_name (TRUE) or feature id (FALSE)}

\item{relative_expr}{A logic flag to determine whether or not we should use the relative expression values}

\item{fitting_type}{The type of fitting procedure used to plot the gene expression dynamics curve, either loess (default) or spline curves.}

\item{log}{A logic flag to determine whether or not you want to log your data (or y-axis is in log scale)}

\item{turning_point}{A logic flag to determine whether or not you want to add two lines indicating the turning point for each gene.}

\item{data_points}{A logic flag to determine whether or not you want to plot the raw data points}
}
\value{
a ggplot2 plot object
}
\description{
This plotting function is used to make kinetic plots for gene pairs along the pseudotime. It accepts a matrix where each 
row is the gene pair and the first column is the hypothetical source or regulator while the second column represents the hypothetical 
target. The name in this matrix should match the name in the gene_short_name column of the cds_subset object. Two vertical lines will 
be added which indicate the estimated inflection points of the two kinetic curves. The distance between the two curves implies 
the potential time delay betweeen two genes. Since Scribe doesn't consider the continous value of the pseudotime but merely the 
ordering of the cells along the pseudotime, the x-axis of this plot also only includes the ordering information. Also note that 
this function allow you to pass a matrix where genes are used across different rows.
}
\examples{
\dontrun{
lung <- load_lung() 
gene_pairs_mat <- matrix(c('H19', 'Ccnd2', 'Ccnd2', 'Scnn1g'), ncol = 2)
plot_gene_pairs_in_pseudotime(lung, gene_pairs_mat)
}
}
