% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_regulation_over_time}
\alias{plot_regulation_over_time}
\title{Function to plot the RDI value over pseudotime (or maybe just a single curve)}
\usage{
plot_regulation_over_time(res, gene_name_vec)
}
\description{
Function to plot the RDI value over pseudotime (or maybe just a single curve)
}
