% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdi_many_runs}
\alias{rdi_many_runs}
\title{rdi_many_runs}
\usage{
rdi_many_runs(x, y, uniformalize)
}
\arguments{
\item{x}{a random variable with multiple run of the same process.}

\item{y}{another random variable with multiple run of another process.}

\item{uniformalize}{Whether or not you want to use ucmi to calculate rdi. Default to be false.}
}
\value{
a numeric value storing the DI from two multiple run variables
}
\description{
This function estimates the DIRECTED mutual information from X to Y when you have multiple run of the processes.
}
\details{
\code{rdi_many_runs} takes two random variables (with the same multiple realization of the two processes) and estimate 
the direct information between them at each time point. It then sums up those information estimators. This function can
only be used when you have hundreds runs of the same time-series experiment.
}
