% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdi_single_run}
\alias{rdi_single_run}
\title{rdi_single_run}
\usage{
rdi_single_run(x, y, d, uniformalize)
}
\arguments{
\item{x}{one random variable from the time-series data}

\item{y}{another random variable from the time-series data}

\item{d}{delay in the formula I(x-->y)=I(x_{t-d};y_t|y_{t-1}) default: d=1}

\item{uniformalize}{Whether or not you want to use ucmi to calculate rdi. Default to be false.}
}
\value{
a numeric value for the restricted direct information between x and y with a time delay d = 1.
}
\description{
This function estimates the RESTRICTED DIRECTED mutual information from X to Y when you have a SINGLE run of the processes
}
\details{
\code{rdi_single_run} takes two random variables x and y as well as a time delay d to estimate the restricted direct infomation between them.
}
