% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdi_single_run_conditioned}
\alias{rdi_single_run_conditioned}
\title{rdi_single_run_conditioned}
\usage{
rdi_single_run_conditioned(x, y, z, z_delays, d, uniformalize)
}
\arguments{
\item{x}{one random variable from the time-series data}

\item{y}{another random variable from the time-series data}

\item{z}{z is a dataframe or matrix consisting of the data for different variables which will be conditioned on.}

\item{z_delays}{z_delay is also a dataframe or matrix consisting of the delays to be applied to different variables}

\item{d}{delay in the formula I(x-->y)=I(x_{t-d};y_t|y_{t-1}) default: d=1}

\item{uniformalize}{Whether or not you want to use ucmi to calculate rdi. Default to be false.}
}
\value{
a numeric value for the estimated condition mutual information between variable x and y conditioning on a third variable z
}
\description{
This function estimates the CONDITIONED DIRECTED mutual information from X to Y CONDITIONED ON Z when you have a SINGLE run of the processes
}
\details{
\code{rdi_single_run_conditioned} takes two random variables x and y as well as the parameter d to calculate the restricted direct information conditioned on variable z.
}
