% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scribe-package.R
\docType{package}
\name{scribe-package}
\alias{scribe}
\alias{scribe-package}
\title{scribe: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jordan Mark Barbone \email{jmbarbone@gmail.com} (\href{https://orcid.org/0000-0001-9788-3628}{ORCID}) [copyright holder]

}
\keyword{internal}
