% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scribe-package.R
\docType{package}
\name{scribe-package}
\alias{scribe}
\alias{scribe-package}
\title{scribe: Command Argument Parsing}
\description{
A base dependency solution with basic argument parsing for use with 'Rscript'.
}
\details{
For a quick overview, see \code{vignette("scribe")}
}
\section{options}{

\describe{
\item{\code{scribe.flag.no}}{\verb{[TRUE]}\cr Default value when \code{scribeArg} uses action \code{"flag"}}
\item{\code{scribe.interactive}}{\verb{[NULL]}\cr Controls \code{interactive()} check within internal \code{exit()}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://jmbarbone.github.io/scribe/}
  \item \url{https://github.com/jmbarbone/scribe}
  \item Report bugs at \url{https://github.com/jmbarbone/scribe/issues}
}

}
\author{
\strong{Maintainer}: Jordan Mark Barbone \email{jmbarbone@gmail.com} (\href{https://orcid.org/0000-0001-9788-3628}{ORCID}) [copyright holder]

}
\keyword{internal}
