% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{smooth_gene}
\alias{smooth_gene}
\title{smooth_gene}
\usage{
smooth_gene(expr_data, window_size = 40L)
}
\arguments{
\item{expr_data}{a matrix for all variables in a time-series where each row is a time point and each column is a variable (for example, a gene). 
The rows are ordered according to time, from earliest to latest.}

\item{window_size}{Integer value for the smoothing window used for calculating the moving average.}
}
\value{
a updated matrix with gene expression smoothed with window size equal to window_size
}
\description{
This subroutine takes a time-series data and returns a moving average for the data.
}
\details{
\code{entropy} takes a integer of dimensions and then calculate the olume of a d-dimensional unit ball for Euclidean norm
using the formula: 0.5 * d * log(pi) - log(gamma(0.5 * d + 1))
It's implimented in C++, providing a (small) increase in speed over the R equivalent.
}
